+++ CHANGELOG.md
@@ -245,3 +245,4 @@
 - GP2-513 - Dropdown-changes
 - noticket bug fix lessons details on objectives
 - GP2-655 - local development support/skeleton for Wagtail Transfer work
+- GP2-667 - appropriately scoped configuration for Wagtail Transfer when deployed